+++ docs/CHANGELOG.rst
@@ -3,6 +3,26 @@ Changelog
 This is a log of the latest changes and improvements to KLibs.
 
 
+0.7.9b1
+-------
+
+(Unreleased)
+
+
+Runtime Changes:
+
+* KLibs now ensures that the participant has been initialized in the database
+  prior to the start of the first block, and likewise ensures that demographics
+  collection is not performed more than once.
+* Demographics collection has been changed so that queries in
+  `user_queries.json` are skipped if they do not correspond to a column in the
+  `participants` table of the database. Additionally, the query for the
+  participant's unique identifer is now always collected first.
+* Demographics collection now fails immediately with an informative message if
+  a required column in the `participants` table does not have a corresponding
+  query.
+
+
 0.7.8b2
 -------